IntelliJ code cleanup

+++ mars/venus/RunSpeedPanel.java
@@ -119,7 +119,6 @@ public class RunSpeedPanel extends JPanel {
      */
     private String setLabel(int index) {
         String result = "Run speed ";
-        ;
         if (index <= SPEED_INDEX_INTERACTION_LIMIT) {
             if (speedTable[index] < 1) {
                 result += speedTable[index];
@@ -142,7 +141,7 @@ public class RunSpeedPanel extends JPanel {
         public void stateChanged(ChangeEvent e) {
             JSlider source = (JSlider) e.getSource();
             if (!source.getValueIsAdjusting()) {
-                runSpeedIndex = (int) source.getValue();
+                runSpeedIndex = source.getValue();
             } else {
                 sliderLabel.setText(setLabel(source.getValue()));
             }
